@@ -1,7 +1,9 @@
 using System;
+using System.Collections.Concurrent;
 using System.IO;
 using Kerberos.NET.Client;
 using Kerberos.NET.Configuration;
+using Kerberos.NET.Credentials;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
@@ -11,44 +13,59 @@ namespace RouteServiceAuth
 {
     
     /// <summary>
-    /// Provides an instance of Kerberos client which is rebuilt if there are any changes to the kerberos options
+    /// Provides lazy creation of Kerberos client tied to credentials using them.
+    /// Clients are rebuilt if there are any changes to the kerberos options
     /// </summary>
     public class KerberosClientProvider
     {
-        private readonly FileSystemWatcher _fileWatcher = new FileSystemWatcher();
-        public KerberosClient Client { get; private set; }
+        private readonly IOptionsMonitor<KerberosOptions> _options;
+        private readonly FileSystemWatcher _fileWatcher = new();
+        ConcurrentDictionary<Credential, KerberosClient> _clientsCache = new();
+        private Krb5Config _krb5config = Krb5Config.Default();
+
 
         public KerberosClientProvider(IOptionsMonitor<KerberosOptions> options)
         {
-            CreateClient(options.CurrentValue);
-            options.OnChange(CreateClient);
+            _options = options;
+            options.OnChange(CreateKrb5Config);
             _fileWatcher.NotifyFilter = NotifyFilters.LastWrite | NotifyFilters.CreationTime;
-            _fileWatcher.Changed += (sender, args) => CreateClient(options.CurrentValue);
+            _fileWatcher.Changed += (sender, args) => CreateKrb5Config(options.CurrentValue);
         }
 
-        private void CreateClient(KerberosOptions options)
+        public KerberosClient GetClient(Credential credential)
         {
-            // var krb5Config = Krb5Config.CurrentUser(options.Krb5ConfigPath);
-            var krb5Config = Krb5Config.Default();
-            if (options.Kdc != null && options.Realm != null)
-            {
-                krb5Config.Realms[options.Realm].Kdc.Add(options.Kdc);
-            }
+            var client = _clientsCache.GetOrAdd(credential, credential => CreateKerberosClient());
+            return client;
+        }
+
+        private void CreateKrb5Config(KerberosOptions options)
+        {
             if (File.Exists(options.Krb5ConfigPath))
             {
+                _krb5config = Krb5Config.Parse(File.ReadAllText(options.Krb5ConfigPath));
                 _fileWatcher.EnableRaisingEvents = true;
-                _fileWatcher.Path = Path.GetDirectoryName(options.Krb5ConfigPath);
-                _fileWatcher.Filter = Path.GetFileName(options.Krb5ConfigPath);
+                _fileWatcher.Path = Path.GetDirectoryName(options.Krb5ConfigPath)!;
+                _fileWatcher.Filter = Path.GetFileName(options.Krb5ConfigPath)!;
             }
             else
             {
                 _fileWatcher.EnableRaisingEvents = false;
             }
-            Client = new KerberosClient(krb5Config);
+
+            _clientsCache.Clear();
         }
 
-
-   
-
+        private KerberosClient CreateKerberosClient()
+        {
+            
+            var options = _options.CurrentValue;
+            var krb5Config = Krb5Config.Parse(_krb5config.Serialize()); // clone
+            if (options.Kdc != null && options.Realm != null)
+            {
+                krb5Config.Realms[options.Realm].Kdc.Add(options.Kdc);
+            }
+            var client = new KerberosClient(krb5Config);
+            return client;
+        }
     }
 }
