@@ -1,6 +1,6 @@
 namespace RouteServiceAuth
 {
-    public class AuthorizationPolicies
+    public static class AuthorizationPolicies
     {
         public const string RequireAuthenticatedUser = "RequireAuthenticatedUser";
     }
