@@ -1,6 +1,6 @@
 namespace RouteServiceAuth
 {
-    public static class Constants
+    public static class KnownHeaders
     {
         public const string X_CF_Forwarded_Url = "X-CF-Forwarded-Url";
         public const string X_CF_Identity = "X-CF-Identity";
