@@ -1,20 +1,23 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Security.Claims;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
 using Kerberos.NET;
+using Kerberos.NET.Crypto;
+using Kerberos.NET.Entities;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Net.Http.Headers;
+using RouteServiceAuth.Kerberos.NET;
 
 namespace RouteServiceAuth
 {
     public class SpnegoAuthenticationHandler : AuthenticationHandler<SpnegoAuthenticationOptions>
     {
-        protected SpnegoAuthenticator _authenticator;
         private const string SchemeName = "Negotiate";
 
 
@@ -24,8 +27,6 @@ namespace RouteServiceAuth
             UrlEncoder encoder,
             ISystemClock clock) : base(optionsMonitor, loggerFactory, encoder, clock)
         {
-            var options = optionsMonitor.CurrentValue;
-            _authenticator = new SpnegoAuthenticator(options);
         }
 
         protected override async Task<AuthenticateResult> HandleAuthenticateAsync()
@@ -48,7 +49,7 @@ namespace RouteServiceAuth
             {
                 Logger.LogTrace("===SPNEGO Token===");
                 Logger.LogTrace(base64Token);
-                var identity = await _authenticator.Authenticate(base64Token);
+                var identity = await GetKerberosAuthenticator().Authenticate(base64Token);
 
                 var ticket = new AuthenticationTicket(
                     new ClaimsPrincipal(identity),
@@ -64,6 +65,30 @@ namespace RouteServiceAuth
             {
                 return AuthenticateResult.Fail(e.ToString());
             }
+        }
+
+        private KerberosAuthenticator GetKerberosAuthenticator()
+        {
+            if(Options.PrincipalName == null)
+                throw new InvalidOperationException(@"Principal name must be in either DOMAIN\account or account@domain format and is case sensitive");
+            if(Options.PrincipalPassword == null)
+                throw new InvalidOperationException(@"Principal password is not specified");
+
+            var principal = KerberosHelper.PrincipalFromUsername(Options.PrincipalName);
+
+            KerberosAuthenticator authenticator;
+            if (Options.PrincipalPassword != null)
+            {
+                var kerberosKey = new KerberosKey(Options.PrincipalPassword, principal, saltType: SaltType.ActiveDirectoryUser);
+                authenticator = new KerberosAuthenticator(new KerberosValidator(kerberosKey));
+            }
+            else
+            {
+                authenticator = new KerberosAuthenticator(new KeyTable(File.ReadAllBytes(Options.KeytabFile)));
+            }
+
+            authenticator.UserNameFormat = UserNameFormat.DownLevelLogonName;
+            return authenticator;
         }
         protected override Task HandleChallengeAsync(AuthenticationProperties properties)
         {
