@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Reflection.Metadata;
 using System.Security.Claims;
 using Kerberos.NET.Client;
 using Kerberos.NET.Configuration;
@@ -9,6 +10,7 @@ using Kerberos.NET.Credentials;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting.Server.Features;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -16,6 +18,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using ProxyKit;
 using RouteServiceAuth.Kerberos.NET;
+using Steeltoe.Common;
 
 namespace RouteServiceAuth
 {
@@ -24,39 +27,38 @@ namespace RouteServiceAuth
         private readonly ILogger<Startup> _logger;
 
         private readonly IConfiguration _configuration;
-        private readonly KerberosProxyOptions _proxySettings;
 
-        public Startup(ILogger<Startup> logger, IConfiguration configuration, IOptions<KerberosProxyOptions> options)
+        public Startup(ILogger<Startup> logger, IConfiguration configuration)
         {
             _logger = logger;
             _configuration = configuration;
-            _proxySettings = options.Value;
         }
 
         public void ConfigureServices(IServiceCollection services)
         {
-            // services.AddOptions<>()
-            services.AddAuthentication().AddJwtBearer().AddSpnegoProxy(_proxySettings, _configuration);
+            services.AddAuthentication()
+                .AddSpnegoProxy();
+            services.AddHttpContextAccessor();
             services.AddAuthorization(cfg =>
             {
-                cfg.AddPolicy(AuthorizationPolicies.RequireAuthenticatedUser, policy => policy.RequireRole().RequireAuthenticatedUser());
+                cfg.AddPolicy(AuthorizationPolicies.RequireAuthenticatedUser, policy => policy.RequireAuthenticatedUser());
             });
 
-            services.Configure<KerberosProxyOptions>(_configuration);
+            services.ConfigureProxyOptions(_configuration);
             services.Configure<KerberosOptions>(_configuration.GetSection("Kerberos"));
             services.Configure<LdapOptions>(_configuration.GetSection("Ldap"));
             services.AddClaimTransformer<LdapRolesClaimsTransformer>();
-            services.AddKerberosClient();
+            services.AddSingleton<KerberosClientProvider>();
             services.AddProxy();
             
         }
         
         public void Configure(IApplicationBuilder app)
         {
-            var ctx = app.ApplicationServices.GetRequiredService<IHttpContextAccessor>();
-            
             app.UseKerberosIngressProxy();
             app.UseKerberosEgressProxy();
+
+            app.MapProxyPorts();
         }
     }
 }
